Make proof contents and verification functionality clearer
The original text said: "When you open `simple_proof.json` in an editor, you will see a JSON file with lots of non-human-readable byte arrays. You can decode this file by running:"

This confused me when I was trying to learn how to use TLSN, because it sounds like the proof does not contain the plain text, but that it rather has some kind of byte-oriented data  "decoded", which sounded to me like it did not contain any direct equivalent to the plaintext—I thought that it must be in encrypted form in the proof. But but the proof does include the plaintext, it's just that it's in byte array form, which is being verified and formatted by the verifier.